split collage into 3 files
